perf: Refactor SearchTypeMenuPopover to improve bottom tabs switching
